(system) refactored folder structure, inserted api folder for publishing DBs to the client
